Bomb

+++ src/uet/oop/bomberman/entities/tile/BrickTile.java
@@ -4,7 +4,7 @@ import uet.oop.bomberman.entities.Entity;
 import uet.oop.bomberman.graphics.Sprite;
 
 public class BrickTile extends Tile {
-    public BrickTile(int x, int y, Sprite sprite) {
+    public BrickTile(double x, double y, Sprite sprite) {
         super(x, y, sprite);
     }
